apps/control-plane/requirements.txt
fastapi>=0.111.0,<1
uvicorn[standard]>=0.30.0,<1
httpx>=0.27.0,<1
pydantic>=2.6.0,<3
PyJWT>=2.8.0,<3
python-jose[cryptography]>=3.3.0,<4
